backend/modules/scraping.py
import requests
from bs4 import BeautifulSoup
import re
from urllib.parse import urlparse

class BrandScraper:
    """Class to scrape brand data from websites"""
    
    def __init__(self):
        self.headers = {
            'User-Agent': 'Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/91.0.4472.124 Safari/537.36'
        }
    
    def scrape_brand_data(self, url):
        """
        Scrape brand data from the provided URL
        
        Args:
            url (str): The brand's website URL
            
        Returns:
            dict: Brand data including name, taglines, product names, descriptions
        """
        # This is a placeholder for the real implementation in Step 4
        return {
            "success": True,
            "message": "Brand scraping placeholder. Will be implemented in Step 4.",
            "brand_data": {
                "name": "Sample Brand",
                "taglines": ["Sample tagline"],
                "products": ["Product 1", "Product 2"],
                "descriptions": ["Product description 1", "Product description 2"]
            }
        } 