backend/requirements.txt
flask==2.0.1
flask-cors==3.0.10
python-dotenv==0.19.0
firebase-admin==5.0.3
requests==2.26.0
beautifulsoup4==4.10.0
lxml==4.6.3
pytest==6.2.5
uuid==1.30 